AuthController, edit login/register BE and FE
