+++ frontend/index.html
@@ -40,6 +40,7 @@
           />
         </div>
 
+        <div class="form-error-message" id="login-error-message"></div>
         <button id="login-button">Login Now</button>
 
         <div class="login-register-switch-text">
@@ -103,6 +104,7 @@
           />
         </div>
 
+        <div class="form-error-message" id="register-error-message"></div>
         <button id="register-button">Register Now</button>
 
         <div class="login-register-switch-text">
