+++ frontend/modules/main.js
@@ -1,14 +1,29 @@
 import { loginUser, logout, registerUser } from "./api.js";
 import { initializeLoginRegister } from "../modules/login-register.js";
-import { validateEmail } from "./helpers.js";
+import { validateEmail, validateName, validatePassword } from "./helpers.js";
 
 document.addEventListener("DOMContentLoaded", () => {
   initializeLoginRegister();
+
+  function displayError(form, message) {
+    const errorDiv = document.getElementById(`${form}-error-message`);
+    if (errorDiv) {
+      errorDiv.innerHTML = message;
+    }
+  }
+
+  function clearError(form) {
+    const errorDiv = document.getElementById(`${form}-error-message`);
+    if (errorDiv) {
+      errorDiv.innerHTML = "";
+    }
+  }
 
   const loginButton = document.getElementById("login-button");
   if (loginButton) {
     loginButton.addEventListener("click", async (event) => {
       event.preventDefault();
+      clearError("login");
 
       const email = document
         .getElementById("login-email")
@@ -17,19 +32,18 @@ document.addEventListener("DOMContentLoaded", () => {
       const password = document.getElementById("login-password").value.trim();
 
       if (!email || !password) {
-        alert("Please fill in all fields.");
+        displayError("login", "Please fill in all fields.");
         return;
       }
 
       if (!validateEmail(email)) {
-        alert("Invalid email format.");
+        displayError("login", "Invalid email format.");
         return;
       }
 
       loginButton.disabled = true;
-      const token = await loginUser(email, password);
 
-      if (!token) {
+      if (!(await loginUser(email, password))) {
         loginButton.disabled = false;
         return;
       }
@@ -48,6 +62,7 @@ document.addEventListener("DOMContentLoaded", () => {
   if (registerButton) {
     registerButton.addEventListener("click", async (event) => {
       event.preventDefault();
+      clearError("register");
 
       const firstName = document
         .getElementById("register-first-name")
@@ -64,19 +79,34 @@ document.addEventListener("DOMContentLoaded", () => {
         .value.trim();
 
       if (!firstName || !lastName || !email || !password) {
-        alert("Please fill in all fields.");
+        displayError("register", "Please fill in all fields.");
+        return;
+      }
+
+      if (!validateName(firstName) || !validateName(lastName)) {
+        displayError(
+          "register",
+          "First and last name must have at least 2 characters."
+        );
         return;
       }
 
       if (!validateEmail(email)) {
-        alert("Invalid email format.");
+        displayError("register", "Invalid email format.");
+        return;
+      }
+
+      if (!validatePassword(password)) {
+        displayError(
+          "register",
+          "Password must be at least 5 characters long and contain a number."
+        );
         return;
       }
 
       registerButton.disabled = true;
-      const token = await registerUser(firstName, lastName, email, password);
 
-      if (!token) {
+      if (!(await registerUser(firstName, lastName, email, password))) {
         registerButton.disabled = false;
         return;
       }
